Struct libp2p_gossipsub::metrics::Metrics

pub struct Metrics { /* private fields */ }

A collection of metrics used throughout the Gossipsub behaviour.

Implementations

impl Metrics

pub fn new(registry: &mut Registry, config: Config) -> Self

pub fn set_topic_peers(&mut self, topic: &TopicHash, count: usize)

Register how many peers do we known are subscribed to this topic.

pub fn joined(&mut self, topic: &TopicHash)

Registers the subscription to a topic if the configured limits allow it. Sets the registered number of peers in the mesh to 0.

pub fn left(&mut self, topic: &TopicHash)

Registers the unsubscription to a topic if the topic was previously allowed. Sets the registered number of peers in the mesh to 0.

pub fn peers_included(
&mut self,
topic: &TopicHash,
reason: Inclusion,
count: usize
)

Register the inclusion of peers in our mesh due to some reason.

pub fn peers_removed(&mut self, topic: &TopicHash, reason: Churn, count: usize)

Register the removal of peers in our mesh due to some reason.

pub fn set_mesh_peers(&mut self, topic: &TopicHash, count: usize)

Register the current number of peers in our mesh for this topic.

pub fn register_invalid_message(&mut self, topic: &TopicHash)

Register that an invalid message was received on a specific topic.

pub fn register_score_penalty(&mut self, penalty: Penalty)

Register a score penalty.

pub fn register_published_message(&mut self, topic: &TopicHash)

Registers that a message was published on a specific topic.

pub fn msg_sent(&mut self, topic: &TopicHash, bytes: usize)

Register sending a message over a topic.

pub fn msg_recvd(&mut self, topic: &TopicHash)

Register that a message was received (and was not a duplicate).

pub fn msg_recvd_unfiltered(&mut self, topic: &TopicHash, bytes: usize)

Register that a message was received (could have been a duplicate).

pub fn register_msg_validation(
&mut self,
topic: &TopicHash,
validation: &MessageAcceptance
)

pub fn memcache_miss(&mut self)

Register a memcache miss.

pub fn register_iwant(&mut self, topic: &TopicHash)

Register sending an IWANT msg for this topic.

pub fn observe_heartbeat_duration(&mut self, millis: u64)

Observes a heartbeat duration.

pub fn observe_mesh_peers_score(&mut self, topic: &TopicHash, score: f64)

Observe a score of a mesh peer.

pub fn peer_protocol_connected(&mut self, kind: PeerKind)

Register a new peers connection based on its protocol.

pub fn peer_protocol_disconnected(&mut self, kind: PeerKind)

Removes a peer from the counter based on its protocol when it disconnects.