Struct libp2p_gossipsub::PeerScoreParams

pub struct PeerScoreParams {
    pub topics: HashMap<TopicHash, TopicScoreParams>,
    pub topic_score_cap: f64,
    pub app_specific_weight: f64,
    pub ip_colocation_factor_weight: f64,
    pub ip_colocation_factor_threshold: f64,
    pub ip_colocation_factor_whitelist: HashSet<IpAddr>,
    pub behaviour_penalty_weight: f64,
    pub behaviour_penalty_threshold: f64,
    pub behaviour_penalty_decay: f64,
    pub decay_interval: Duration,
    pub decay_to_zero: f64,
    pub retain_score: Duration,
}

Fields
topics: HashMap<TopicHash, TopicScoreParams>

Score parameters per topic.

topic_score_cap: f64

Aggregate topic score cap; this limits the total contribution of topics towards a positive score. It must be positive (or 0 for no cap).

app_specific_weight: f64

P5: Application-specific peer scoring

ip_colocation_factor_weight: f64

P6: IP-colocation factor. The parameter has an associated counter which counts the number of peers with the same IP. If the number of peers in the same IP exceeds ip_colocation_factor_threshold, then the value is the square of the difference, ie (peers_in_same_ip - ip_colocation_threshold)^2`. If the number of peers in the same IP is less than the threshold, then the value is 0. The weight of the parameter MUST be negative, unless you want to disable for testing. Note: In order to simulate many IPs in a manageable manner when testing, you can set the weight to 0 thus disabling the IP colocation penalty.

ip_colocation_factor_threshold: f64ip_colocation_factor_whitelist: HashSet<IpAddr>behaviour_penalty_weight: f64

P7: behavioural pattern penalties. This parameter has an associated counter which tracks misbehaviour as detected by the router. The router currently applies penalties for the following behaviors:

attempting to re-graft before the prune backoff time has elapsed.

not following up in IWANT requests for messages advertised with IHAVE.

The value of the parameter is the square of the counter over the threshold, which decays with BehaviourPenaltyDecay. The weight of the parameter MUST be negative (or zero to disable).

behaviour_penalty_threshold: f64behaviour_penalty_decay: f64decay_interval: Duration

The decay interval for parameter counters.

decay_to_zero: f64

Counter value below which it is considered 0.

retain_score: Duration

Time to remember counters for a disconnected peer.

Implementations

impl PeerScoreParams

Peer score parameter validation

pub fn validate(&self) -> Result<(), String>