Crate libp2p_gossipsub

Implementation of the Gossipsub protocol.

Gossipsub is a P2P pubsub (publish/subscription) routing layer designed to extend upon floodsub and meshsub routing protocols.

Overview

Note: The gossipsub protocol specifications (https://github.com/libp2p/specs/tree/master/pubsub/gossipsub) provide an outline for the routing protocol. They should be consulted for further detail.

Gossipsub is a blend of meshsub for data and randomsub for mesh metadata. It provides bounded degree and amplification factor with the meshsub construction and augments it using gossip propagation of metadata with the randomsub technique.

The router maintains an overlay mesh network of peers on which to efficiently send messages and metadata. Peers use control messages to broadcast and request known messages and subscribe/unsubscribe from topics in the mesh network.

Important Discrepancies

This section outlines the current implementation’s potential discrepancies from that of other implementations, due to undefined elements in the current specification.

Topics - In gossipsub, topics configurable by the hash_topics configuration parameter. Topics are of type TopicHash. The current go implementation uses raw utf-8 strings, and this is default configuration in rust-libp2p. Topics can be hashed (SHA256 hashed then base64 encoded) by setting the hash_topics configuration parameter to true.

Sequence Numbers - A message on the gossipsub network is identified by the source PeerId and a nonce (sequence number) of the message. The sequence numbers in this implementation are sent as raw bytes across the wire. They are 64-bit big-endian unsigned integers. When messages are signed, they are monotonically increasing integers starting from a random value and wrapping around u64::MAX. When messages are unsigned, they are chosen at random. NOTE: These numbers are sequential in the current go implementation.

Peer Discovery

Gossipsub does not provide peer discovery by itself. Peer discovery is the process by which peers in a p2p network exchange information about each other among other reasons to become resistant against the failure or replacement of the boot nodes of the network.

Peer discovery can e.g. be implemented with the help of the Kademlia protocol in combination with the Identify protocol. See the Kademlia implementation documentation for more information.

Using Gossipsub

Gossipsub Config

The Config struct specifies various network performance/tuning configuration parameters. Specifically it specifies:

This struct implements the Default trait and can be initialised via Config::default().

Behaviour

The Behaviour struct implements the libp2p_swarm::NetworkBehaviour trait allowing it to act as the routing behaviour in a libp2p_swarm::Swarm. This struct requires an instance of PeerId and Config.

Example

For an example on how to use gossipsub, see the chat-example.

Structs

AllowAllSubscriptionFilter

Allows all subscriptions

Behaviour

Network behaviour that handles the gossipsub protocol.

CallbackSubscriptionFilter

CombinedSubscriptionFilters

Combines two subscription filters

Config

Configuration parameters that define the performance of the gossipsub network.

ConfigBuilder

The builder struct for constructing a gossipsub configuration.

IdentityTransform

The default transform, the raw data is propagated as is to the application layer gossipsub.

MaxCountSubscriptionFilter

Adds a max count to a given subscription filter

Message

The message sent to the user after a RawMessage has been transformed by a crate::DataTransform.

MessageId

MetricsConfig

PeerScoreParams

PeerScoreThresholds

RawMessage

A message received by the gossipsub system and stored locally in caches..

RegexSubscriptionFilter

A subscription filter that filters topics based on a regular expression.

Topic

A gossipsub topic.

TopicHash

TopicScoreParams

WhitelistSubscriptionFilter

Allows only whitelisted subscriptions

Enums

ConfigBuilderError

Error associated with Config building.

Event

Event that can be emitted by the gossipsub behaviour.

MessageAcceptance

Validation kinds from the application for received messages.

MessageAuthenticity

Determines if published messages should be signed or not.

PublishError

Error associated with publishing a gossipsub message.

SubscriptionError

Error associated with subscribing to a topic.

ValidationError

ValidationMode

The types of message validation that can be employed by gossipsub.

Version

Selector for custom Protocol Id

Traits

DataTransform

A general trait of transforming a RawMessage into a Message. The RawMessage is obtained from the wire and the Message is used to calculate the crate::MessageId of the message and is what is sent to the application.

Hasher

A generic trait that can be extended for various hashing types for a topic.

TopicSubscriptionFilter

Functions

score_parameter_decay

Computes the decay factor for a parameter, assuming the decay_interval is 1s and that the value decays to zero if it drops below 0.01.

score_parameter_decay_with_base

Computes the decay factor for a parameter using base as the decay_interval.

Type Aliases

IdentTopic

RpcDeprecated

Sha256Topic