Trait libp2p_gossipsub::DataTransform

pub trait DataTransform {
    fn inbound_transform(
        &self,
        raw_message: RawGossipsubMessage
    ) -> Result<GossipsubMessage, Error>;
    fn outbound_transform(
        &self,
        topic: &TopicHash,
        data: Vec<u8>
    ) -> Result<Vec<u8>, Error>;
}

A general trait of transforming a RawGossipsubMessage into a GossipsubMessage. The RawGossipsubMessage is obtained from the wire and the GossipsubMessage is used to calculate the crate::MessageId of the message and is what is sent to the application.

The inbound/outbound transforms must be inverses. Applying the inbound transform and then the outbound transform MUST leave the underlying data un-modified.

By default, this is the identity transform for all fields in GossipsubMessage.

Required Methods

fn inbound_transform(
&self,
raw_message: RawGossipsubMessage
) -> Result<GossipsubMessage, Error>

Takes a RawGossipsubMessage received and converts it to a GossipsubMessage.

fn outbound_transform(
&self,
topic: &TopicHash,
data: Vec<u8>
) -> Result<Vec<u8>, Error>

Takes the data to be published (a topic and associated data) transforms the data. The transformed data will then be used to create a crate::RawGossipsubMessage to be sent to peers.